bup-gc: also cleanup par2-files
Do not abort, if the to-be-cleaned-up file does not exist (f.e. non-existant par2-files)

+++ lib/bup/gc.py
@@ -137,6 +137,8 @@ def sweep(live_objects, existing_count, cat_pipe, threshold, compression,
         for p in ns.stale_files:
             if new_pack_prefix and p.startswith(new_pack_prefix):
                 continue # Don't remove the new pack file
+            if not os.access(p, os.F_OK): # File does not exist (f.e. n/e par2-file)
+                continue
             if verbosity:
                 log('removing ' + basename(p) + '\n')
             os.unlink(p)
@@ -170,6 +172,9 @@ def sweep(live_objects, existing_count, cat_pipe, threshold, compression,
                     % git.repo_rel(basename(idx_name)))
             ns.stale_files.append(idx_name)
             ns.stale_files.append(idx_name[:-3] + 'pack')
+            # FIXME: other par2 sizes? Use glob instead?
+            ns.stale_files.append(idx_name[:-3] + 'par2')
+            ns.stale_files.append(idx_name[:-3] + 'vol000+200.par2')
             continue
 
         live_frac = idx_live_count / float(len(idx))
@@ -191,6 +196,9 @@ def sweep(live_objects, existing_count, cat_pipe, threshold, compression,
 
         ns.stale_files.append(idx_name)
         ns.stale_files.append(idx_name[:-3] + 'pack')
+        # FIXME: other par2 sizes? Use glob instead?
+        ns.stale_files.append(idx_name[:-3] + 'par2')
+        ns.stale_files.append(idx_name[:-3] + 'vol000+200.par2')
 
     if verbosity:
         progress('preserving live data (%d%% complete)\n'
